Crea test.

+++ src/main/java/grupo3/clase7/comercioInformatico/ServicioDeArmado.java
@@ -1,14 +1,14 @@
 package grupo3.clase7.comercioInformatico;
 
 public class ServicioDeArmado extends Servicio{
-	public final double VALOR_HORA = 250;
+	public final int VALOR_HORA = 250;
 	
 	public ServicioDeArmado(int cantidadHoras) {
 		super (cantidadHoras);
 	}
 
 	@Override
-	public double getValorHora() {
+	public int getValorHora() {
 		return this.VALOR_HORA;
 	}
 }
